% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleplot.R
\name{airflow_task_state_colors}
\alias{airflow_task_state_colors}
\title{Airflow task-state colors}
\usage{
airflow_task_state_colors()
}
\value{
data frame of the form (state, color)
}
\description{
Convenience function to get state-color mappings for Airflow
}
