% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_by_distance.R
\name{bin_by_distance}
\alias{bin_by_distance}
\title{Bin Data by Distance}
\usage{
bin_by_distance(
  data,
  distance_col = "Y_um",
  bin_width = 10,
  groups = c("Species", "Animal_ID", "Position", "Section_No", "Column"),
  min_col = "VZ_base_um",
  max_col = "Cortex_top_um",
  suffix = "_bin"
)
}
\arguments{
\item{data}{A dataframe containing the data to be binned.}

\item{distance_col}{The name of the column containing distance values (default: 'Y_um').}

\item{bin_width}{The width of each bin (in the same unit as `distance_col`).}

\item{groups}{A character vector specifying grouping columns for binning (default: c("Species", "Animal_ID", "Position", "Section_No", "Column")).}

\item{min_col}{The name of the column containing the minimum value for binning (default: 'VZ_base_um').}

\item{max_col}{The name of the column containing the maximum value for binning (default: 'Cortex_top_um').}

\item{suffix}{The suffix to append to the new binned distance column name (default: "_bin").}
}
\value{
A dataframe with an additional column representing binned distance values.
}
\description{
Bins distance data based on specified parameters and appends a cleaned-up binned distance column to the dataframe.
}
\details{
This function bins distance values from the specified column into discrete intervals (bins) defined by the `bin_width` parameter,
grouped by specified columns (`groups`). The resulting binned distances are added as a new column with a suffix (`suffix`) to
the input dataframe (`data`).
}
\examples{
\dontrun{
# Bin data by distance with default parameters
bin_by_distance(data, bin_width = 10)
}

}
