% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleplot.R
\name{circleplot}
\alias{circleplot}
\title{Circle Plot}
\usage{
circleplot(id, states, colors, width = NULL, height = NULL)
}
\arguments{
\item{id}{id assigned to the html element}

\item{states}{data frame of the form (state, count, url).}

\item{colors}{data frame of the form (state, color). All states will be displayed. Row order determines display order.}
}
\value{
r2d3 htmlwidget
}
\description{
Function to create a d3-based horizontal circle plot
}
\examples{
 \dontrun{circleplot('af_plot', 1:8, airflow_task_state_colors())}
}
