% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_bin_labels.R
\name{cleanup_bin_labels}
\alias{cleanup_bin_labels}
\title{This function cleans up labels from binning using `cut`. It uses `gsub` to do this.}
\usage{
cleanup_bin_labels(labels)
}
\arguments{
\item{labels}{A set of labels produced by cut}
}
\value{
The same labels but formatted as "lower - upper" without additional brackets.
}
\description{
This function cleans up labels from binning using `cut`. It uses `gsub` to do this.
}
\examples{
# Example usage
labels <- cleanup_bin_labels(labels)

# If replacing levels of a factor, as in `bin_by_distance`

levels(factor) <- cleanup_bin_labels(levels(factor))

}
