% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_distance_columns.R
\name{clip_distance_columns}
\alias{clip_distance_columns}
\title{Clip Distance Columns to Limits}
\usage{
clip_distance_columns(
  df,
  distance_columns,
  min_col = "VZ_base",
  max_col = "Cortex_top"
)
}
\arguments{
\item{df}{A dataframe.}

\item{distance_columns}{Character vector of column names containing distance data.}

\item{min_col}{A string that matches a column name containing the lower bounds to clip to (default: "VZ_base").}

\item{max_col}{A string that matches a column name containing the upper bounds to clip to (default: "Cortex_top").}
}
\value{
A dataframe with clipped distance columns.
}
\description{
This function clips distance columns within a specified range defined by min_col and max_col.
It is a wrapper function that applies the \code{\link{clip_to_limits}} function to each distance column.
}
\examples{
clip_distance_columns(df = counts_with_sections_data,
                       distance_columns = c("dist_col1", "dist_col2"),
                       min_col = "VZ_base",
                       max_col = "Cortex_top")

}
\seealso{
\code{\link{clip_to_limits}}: The underlying function used for clipping individual distance columns.
}
