% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_to_limits.R
\name{clip_to_limits}
\alias{clip_to_limits}
\title{Clip Values to Limits}
\usage{
clip_to_limits(
  df,
  distance_col,
  min_limit_col = "VZ_base",
  max_limit_col = "Cortex_top"
)
}
\arguments{
\item{df}{A dataframe.}

\item{distance_col}{A string that matches a column name containing distance data.}

\item{min_limit_col}{A string that matches a column name containing the lower bounds to clip to (default: "VZ_base").}

\item{max_limit_col}{A string that matches a column name containing the upper bounds to clip to (default: "Cortex_top").}
}
\value{
A numeric vector of equal length to distance_col.
}
\description{
This function clips values within a specified range for a given distance column in a dataframe.
}
\examples{
clip_to_limits(df = counts_with_sections_data,
               distance_col = "your_distance_column",
               min_limit_col = "your_min_limit_column",
               max_limit_col = "your_max_limit_column")

}
