% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_data_dirs.R
\name{construct_data_dirs}
\alias{construct_data_dirs}
\title{Construct Data Directories}
\usage{
construct_data_dirs(
  data_dir = NULL,
  ep = NULL,
  collect = NULL,
  exp_con = NULL,
  ctl_con = NULL,
  assay = NULL
)
}
\arguments{
\item{data_dir}{Vector of directory paths (default: NULL)}

\item{ep}{Experimental stage (default: NULL)}

\item{collect}{Collection stage (default: NULL)}

\item{exp_con}{Experimental construct (default: NULL)}

\item{ctl_con}{Control construct (default: NULL)}

\item{assay}{IF assay (default: NULL)}
}
\value{
A vector containing the paths for experimental and control directories
}
\description{
This function constructs the paths for data directories based on the provided or default parameters.
}
\examples{
# Example 1: Using default parameters
construct_data_dirs()

# Example 2: Providing custom parameters
custom_dirs <- construct_data_dirs(data_dirs = c("exp_path", "control_path"), ep = "Stage1", c = "CollectionA")

}
