% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_dataframes.R
\name{construct_dataframes}
\alias{construct_dataframes}
\title{Construct Dataframes}
\usage{
construct_dataframes(
  data_dirs,
  data_type = c("counts", "sections", "conditions")
)
}
\arguments{
\item{data_dirs}{Vector of directory paths}

\item{data_type}{Type of data: 'counts', 'sections', 'conditions' or 'fluorescence'. (default: c('counts','sections','conditions'))

Each of the datatypes should be stored in a .csv file of the same name.  Custom names may be used for these files.}
}
\value{
A dataframe with appended data from specified directories
}
\description{
This function reads and appends data from multiple directories based on the specified data type.  The datatype matches the name of the .csv file.
}
\examples{
# Example:
data_dirs <- c("exp_path", "control_path")
construct_dataframes(data_dirs, data_type = "counts")

}
