% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_percentage.R
\name{convert_to_percentage}
\alias{convert_to_percentage}
\title{Convert Distance to Percentage}
\usage{
convert_to_percentage(
  df,
  distance_col = "Y",
  max_col = "Cortex_top",
  min_col = "VZ_base",
  suffix = "_percent",
  prefix = ""
)
}
\arguments{
\item{df}{A dataframe.}

\item{distance_col}{The column(s) containing individual distance values to be converted (default: 'Y').}

\item{max_col}{The column containing individual maximum value (default: 'Cortex_top').}

\item{min_col}{The column containing individual minimum value (default: 'VZ_base').}

\item{suffix}{The suffix to append to the new percentage column name (default: '_percent').}

\item{prefix}{The prefix to prepend to the new percentage column name (default: "").}
}
\value{
A dataframe with new columns representing the individual distance values converted to a percentage scale.
}
\description{
This function converts individual distance values to a percentage scale based on individual minimum and maximum values.
}
\examples{
# Example usage with multiple distance columns
convert_to_percentage(df = my_data, distance_col = c("dist_col1", "dist_col2"), max_col = "Cortex_top", min_col = "VZ_base", suffix = "_scaled", prefix = "scaled_")

}
