% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_summaries.R
\name{count_summaries}
\alias{count_summaries}
\title{Cell Count Summaries or Densities}
\usage{
count_summaries(
  df,
  summary_type,
  width = NULL,
  height = NULL,
  groups = c("Species", "Construct", "Animal_ID", "Section_No", "Position", "Column"),
  marker = "Marker",
  compartments = "Region"
)
}
\arguments{
\item{df}{A dataframe containing cell data in long format.}

\item{summary_type}{A character string specifying the type of summary ("counts" or "density").}

\item{width}{Name of the column containing cell width measurements for density calculation.}

\item{height}{Name of the column containing cell height measurements for density calculation.}

\item{groups}{Character vector specifying grouping variables for summarisation.}

\item{marker}{Name of the column containing marker information.}

\item{compartments}{Name of the column containing region information.}
}
\value{
A dataframe with count summaries or cell densities based on the specified summary_type.
}
\description{
Computes count summaries or cell densities from a dataframe of cells in long format, based on specified parameters.
}
\details{
The function computes either counts per marker per group or cell densities per marker within specified regions,
depending on the summary_type parameter.

For "counts" summary_type, it calculates the number of cells (counts) per marker per group.
For "density" summary_type, it calculates cell densities (cells per unit area) per marker within specified regions,
using the provided width and height measurements.
}
\examples{
\dontrun{
# Example usage:
df <- data.frame(
  Species = c("Mouse", "Mouse", "Dunnart", "Dunnart"),
  Construct = c("A", "B", "A", "B"),
  Animal_ID = c(1, 2, 1, 2),
  Section_No = c(1, 2, 1, 2),
  Position = c("Lateral", "Dorsal", "Lateral", "Dorsal"),
  Column = c("A1", "B1", "A2", "B2"),
  Marker = c("tdT", "tdT", "Tbr2", "Tbr2"),
  Region = c("VZ", "VZ", "SVZ", "SVZ"),
  Width = c(100.15, 99.55, 100.56, 101.23),
  Height = c(300, 500, 200, 300)
)

# Compute counts per marker per group:
count_summaries(df, summary_type = "counts")

# Compute cell densities per marker within specified regions using width and height measurements:
count_summaries(df, summary_type = "density", width = "Width", height = "Height")
}

}
