% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_filename_variables.R,
%   R/get_channel_columns.R, R/secondary_channels.R
\name{extract_filename_variables}
\alias{extract_filename_variables}
\alias{get_channel_columns}
\alias{secondary_columns}
\title{Extract Filename Variables}
\usage{
extract_filename_variables(
  df,
  species = NULL,
  nuclear_stain = NULL,
  channels = NULL
)

get_channel_columns(column_name, df)

secondary_columns(channels = NULL)
}
\arguments{
\item{df}{The input dataframe.}

\item{species}{A vector containing the species identifier strings, e.g. c("FTD", "Mm")}

\item{nuclear_stain}{A string containing the name of the nuclear stain ,e.g. DAPI}

\item{channels}{The secondary channels.}

\item{column_name}{The name of the column.}
}
\value{
A modified dataframe with added columns.

A modified dataframe with the extracted channel column.

A character vector of channel names.
}
\description{
This function extracts various variables from the 'File' column of a dataframe and adds them as new columns.

This function extracts channel columns based on the provided column name and dataframe.

This function generates channel names based on the provided secondary channels.
}
\examples{
df <- extract_filename_variables(df)

}
