% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_file_templates.R
\name{generate_data_file_templates}
\alias{generate_data_file_templates}
\title{Generate CSV files}
\usage{
generate_data_file_templates(directory, layer_cols = character())
}
\arguments{
\item{directory}{The directory where the .csv files should be generated.}

\item{layer_cols}{A character vector of additional column names to include in the "sections.csv" file, situated between the 'VZ_base' and 'Cortex_top' columns. Default is an empty vector.}
}
\value{
None
}
\description{
This function generates three .csv files in a user-specified directory. These files are in the format used for immunofluorescence data analysed using
this package.

The first file, "counts.csv", has columns 'File', 'Marker', 'X', 'Y', and 'Mean_Gray'.

The second file, "sections.csv", has columns 'File', 'VZ_base', 'Cortex_top', 'Microscope', 'Objective', and 'AgR'. This file can be customized with additional layer columns.

The third file, "conditions.csv", has columns 'Mother_ID', 'Offspring', 'Condition', 'EP_Age', 'Collection_Age', 'EP_Stage', 'Collection_Stage', and 'Construct'.
}
