% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airflow.R
\name{get_airflow_dag_args}
\alias{get_airflow_dag_args}
\title{Get airflow dag args}
\usage{

  get_airflow_dag_args(dag_dir = "/Users/misha.lisovich/Documents/bitbucket/warp-pipelines/warp_pipelines/dags",
  args = c("start_date", "schedule_interval"))
}
\arguments{
\item{dag_dir}{airflow dag directory}

\item{args}{attributes to extract. NOTE: currently only extracts start_date and schedule_interval - needs generalization.}
}
\value{
data frame of the form (dag_id, dag_args)
}
\description{
Get specified args (attributes) from all Airflow dags contained in dag_dir
}
