% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airflow.R
\name{get_csrf_token}
\alias{get_csrf_token}
\title{Get Airflow csrf token}
\usage{
get_csrf_token(airflow_url)
}
\arguments{
\item{airflow_url}{base url for airflow instance}
}
\value{
csrf token string
}
\description{
Get csrf token by scraping Airflow's queryview page
}
