% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hier_conditions.R
\name{get_hier_conditions}
\alias{get_hier_conditions}
\title{Extract HIER Conditions from a DataFrame}
\usage{
get_hier_conditions(df, hier_column = "AgR")
}
\arguments{
\item{df}{The input dataframe.}

\item{hier_column}{The name of the column containing HIER conditions (default: "AgR").}
}
\value{
A modified dataframe with added columns.
}
\description{
This function adds three new columns based on HIER conditions recorded in formatted text from a specified column
The format of the HIER conditions should be as follows:
}
\details{
Molarity-Buffer_TempC_Timemins

For example, using a 0.01 M Sodium Citrate buffer at 110C for 4 mins would equate to:

   0.01-SC_110C_4min

These data fields will then be used to generate separate variable columns.

Options for buffers include:

   SC = Sodium Citrate
   THCl = Tris-HCl
   TEDTA = Tris-EDTA

At present, buffer pH is not included as a field
}
\examples{
sections <- get_hier_conditions(sections)

}
