% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{glicon}
\alias{glicon}
\title{Glyph Icon}
\usage{
glicon(name, class = NULL)
}
\arguments{
\item{name}{Name of icon. Icons are drawn from the
\href{{https://fontawesome.com/}{Font Awesome Free}} (currently icons from
the v5.3.1 set are supported with the v4 naming convention) and
\href{{http://getbootstrap.com/components/#glyphicons}{Glyphicons}}
libraries. Note that the "fa-" and "glyphicon-" prefixes should not be used
in icon names (i.e. the "fa-calendar" icon should be referred to as
"calendar")}

\item{class}{Additional classes to customize the style of the icon (see the
\href{{http://fontawesome.io/examples/}{usage examples}} for details on
supported styles).}
}
\description{
Specialization of shiny::icon using glyphicon library.
}
