% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laminaR-package.R
\docType{package}
\name{laminaR-package}
\alias{laminaR}
\alias{laminaR-package}
\title{laminaR: Utilities for cortical column immunofluorescence data}
\description{
A collection of basic data manipulation functions for use with immunofluorescence count data in cortical columns. The functions in this package were developed for use in the paper _____, and are included here in this package for replicability, but are not actively maintained. This workflow assumes data collected in a specific format across multiple .csv files. These files may be generated by functions within the package, or by accessing the data included in the package (see the associated paper for details).
}
\keyword{internal}
