% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laminar.R
\name{launch_application}
\alias{launch_application}
\title{Launch Laminar}
\usage{
launch_application(appDir = system.file("laminar_app", package =
  "laminar"), ...)
}
\arguments{
\item{appDir}{the application to run. Defaults to application within the installed laminar package.}

\item{...}{any other arguments passed to shiny::runApp}
}
\description{
Launch the Laminar application
}
