% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_fluorescence_by_background.R
\name{normalise_fluorescence_by_background}
\alias{normalise_fluorescence_by_background}
\title{Normalise Fluorescence by Background Value}
\usage{
normalise_fluorescence_by_background(
  df,
  fluor_col = "Gray_Value",
  F0 = "BG",
  marker = "Marker",
  groups = NULL,
  col.name = "deltaF.F0"
)
}
\arguments{
\item{df}{The input dataframe containing fluorescence values in tidy (long) format.}

\item{fluor_col}{The name of the column containing fluorescence values (default: "Gray_Value").}

\item{F0}{The name of the column containing background fluorescence values, "min" to calculate background as the minimum fluorescence per sample group,
or min.max to calculate the min-max normalisation by the sample group (default: "BG").}

\item{marker}{The name of the column containing the channel information (a factor with levels for each marker) (default: "Marker").}

\item{groups}{Grouping variables to specify sample-level groupings for normalisation. Required only if `F0 = "min"` or if `F0 = "min.max"`}

\item{col.name}{Tha name of the resulting normalised fluorescence column (default: "deltaF.F0")}
}
\value{
A dataframe with an additional column for normalised fluorescence values.
}
\description{
This function normalises fluorescence values by subtracting background fluorescence values from a specified column.
The background fluorescence value can be a measured value, calculated as the minimum fluorescence value within a sample group,
or calculated via min-max normalisation based upon minimum and maximum measures in the sample group
}
\examples{
\dontrun{
# Specify columns containing measured data for background fluorescence.
normalise_fluorescence_by_background(df, fluor_col = "Gray_Value", F0 = "BG", marker = "Marker")

# Calculate background fluorescence as the sample minimum with grouping columns specified.
normalise_fluorescence_by_background(df, fluor_col = "Gray_Value", F0 = "min", marker = "Marker", groups = c("Sample_ID"))

#' # Calculate background fluorescence as the sample minimum with grouping columns specified.
normalise_fluorescence_by_background(df, fluor_col = "Gray_Value", F0 = "min.max", marker = "Marker", groups = c("Sample_ID"))
}
}
