% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{reactivePoll2}
\alias{reactivePoll2}
\title{Reactive Polling}
\usage{
reactivePoll2(intervalMillis, session, checkFunc)
}
\arguments{
\item{intervalMillis}{Approximate number of milliseconds to wait between
calls to \code{checkFunc}. This can be either a numeric value, or a
function that returns a numeric value.}

\item{session}{The user session to associate this file reader with, or
\code{NULL} if none. If non-null, the reader will automatically stop when
the session ends.}

\item{checkFunc}{A relatively cheap function whose values over time will be
tested for equality; inequality indicates that the underlying value has
changed and needs to be invalidated and re-read using \code{valueFunc}. See
Details.}
}
\description{
Specialization of shiny::reactivePoll where checkFunc is reused as valueFunc.
Useful making database table sources reactive without querying twice.
}
\examples{
 \dontrun{cars <- reactivePoll(1000, session, collect(tbl(db, sql('select * from cars'))) )}
}
