% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_append_data.R
\name{read_and_append_data}
\alias{read_and_append_data}
\title{Read and Append Data}
\usage{
read_and_append_data(
  assay = NULL,
  data_dirs = NULL,
  ep = NULL,
  collect = NULL,
  exp_con = NULL,
  ctl_con = NULL,
  data_types = c("counts", "sections", "conditions")
)
}
\arguments{
\item{assay}{IF assay (default: NULL)}

\item{data_dirs}{Vector of directory paths for custom data (default: NULL)}

\item{ep}{Experimental stage (default: NULL)}

\item{collect}{Collection stage (default: NULL)}

\item{exp_con}{Experimental construct (default: NULL)}

\item{ctl_con}{Control construct (default: NULL)}

\item{data_types}{A vector containing the type of data that is being imported. Options are 'counts', 'sections', 'conditions' or 'fluorescence',
but can be modified to match custom names of the .csv files containing the data (default: c("counts","sections","conditions"))}
}
\value{
A list of dataframes for the specified datasets ('counts','sections','conditions','fluorescence')
}
\description{
This function acts as a wrapper, allowing the user to specify an assay and importing data for 'counts', 'sections', and 'conditions'.
}
\examples{
# Example 1: Using default parameters
result <- read_and_append_data(assay = "PH3-Tbr2-tdT")

# Example 2: Providing custom parameters
custom_result <- read_and_append_data(assay = "CustomAssay", data_dirs = c("exp_path1", "exp_path2"), ep = "Stage1", c = "CollectionA")

}
