% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_distance.R
\name{rescale_distance}
\alias{rescale_distance}
\title{Rescale Distance Measures}
\usage{
rescale_distance(
  df,
  distance_cols = distance_columns,
  min_col = "VZ_base",
  suffix = "_um"
)
}
\arguments{
\item{df}{A dataframe.}

\item{distance_cols}{A vector containing strings that correspond to columns containing distance data (default: global variable 'distance_columns').}

\item{min_col}{A string that matches the column name of the dataframe column containing the lower bounds of the range (default: 'VZ_base').}

\item{suffix}{A string to append to the end of the names of the output vectors/columns in the dataframe (default: '_um').}
}
\value{
A dataframe with additional modified distance columns/numeric vectors appended.
}
\description{
This function rescales distance measures so that 0 corresponds to a pre-specified value.
In the context of column-based IF analysis in cortical sections, this often corresponds to the ventricular
zone, VZ, such that the transformed distance measures are given as 'distance relative to the VZ'.
}
\examples{
# Example with custom columns and suffix
rescale_distance(df = my_data, distance_cols = c("dist_col1", "dist_col2"), min_col = "Base_value", suffix = "_scaled")

}
